Move upload/edit object view to from users to core
