+++ src/core/tests/test_upload_object_view.py
