+++ src/core/urls.py
@@ -15,10 +15,12 @@ from core.views.views import (
     collection,
     create_exhibit,
     edit_exhibit,
+    edit_object,
     exhibit,
     exhibit_detail,
     exhibit_select,
     manifest,
+    object_upload,
     related_content,
     see_all,
     service_worker,
@@ -34,6 +36,8 @@ urlpatterns = [
     path("exhibits/create/", create_exhibit, name="create-exhibit"),
     path("exhibits/edit/", edit_exhibit, name="edit-exhibit"),
     path("artwork/", artwork_preview, name="artwork-preview"),
+    path("objects/upload/", object_upload, name="object-upload"),
+    path("objects/edit/", edit_object, name="edit-object"),
     path("generator/", marker_generator, name="marker-generator"),
     path("related-content", related_content, name="related-content"),
     path("sw.js", service_worker, name="sw"),
