+++ src/users/urls.py
@@ -8,11 +8,9 @@ from .views import (
     delete,
     edit_artwork,
     edit_marker,
-    edit_object,
     edit_password,
     edit_profile,
     marker_upload,
-    object_upload,
     profile,
     signup,
 )
@@ -47,8 +45,6 @@ urlpatterns = [
     path("profile/edit/", edit_profile, name="edit-profile"),
     path("profile/edit-password/", edit_password, name="edit-password"),
     path("markers/upload/", marker_upload, name="marker-upload"),
-    path("objects/upload/", object_upload, name="object-upload"),
-    path("objects/edit/", edit_object, name="edit-object"),
     path("markers/edit/", edit_marker, name="edit-marker"),
     path("artworks/create/", create_artwork, name="create-artwork"),
     path("artworks/edit/", edit_artwork, name="edit-artwork"),
