+++ src/users/views.py
@@ -16,6 +16,7 @@ from django.urls import reverse_lazy
 from django.utils.translation import gettext_lazy as _
 from django.views.decorators.http import require_http_methods
 
+from core.forms import UploadObjectForm
 from core.models import Artwork, Exhibit, Marker, Object
 
 from .forms import (
@@ -24,7 +25,6 @@ from .forms import (
     ProfileForm,
     SignupForm,
     UploadMarkerForm,
-    UploadObjectForm,
 )
 from .models import Profile
 from .services import BOT_SCORE, create_assessment
@@ -228,11 +228,6 @@ def upload_elements(request, form_class, form_type, route):
 
 
 @login_required
-def object_upload(request):
-    return upload_elements(request, UploadObjectForm, "object", "object-upload")
-
-
-@login_required
 def marker_upload(request):
     return upload_elements(request, UploadMarkerForm, "marker", "marker-upload")
 
@@ -258,29 +253,6 @@ def edit_elements(request, form_class, route, model, model_data):
             "form": form_class(initial=model_data),
             "model": model,
         },
-    )
-
-
-@login_required
-def edit_object(request):
-    index = request.GET.get("id", "-1")
-    model = Object.objects.get(id=index)
-
-    model_data = {
-        "source": model.source,
-        "uploaded_at": model.uploaded_at,
-        "author": model.author,
-        "scale": model.scale,
-        "position": model.position,
-        "rotation": model.rotation,
-        "title": model.title,
-    }
-    return edit_elements(
-        request,
-        UploadObjectForm,
-        route="users/edit-object.jinja2",
-        model=model,
-        model_data=model_data,
     )
